add HiGal module
more cleanup

add changelog

it is possible for a cutout to be missing; we need to allow for that.
Also, the hack I had before isn't needed when you appropriately query
for multiple wavelengths

remove some comments, still not done with the PR

minor correction

added basic docs

pep8 fixes.  The bracket closures DO match the indentation, though.
pep8 lies sometimes.

add local tests and cleanup some testing

more interactive tests revealed that my hacks didn't really work, this
commit has some fixes

move everything into astroquery.herschel.higal

update relative paths

whitespace

update tests

the higal server genuinely isn't showing anything when I search online now

add HiGal module

more cleanup

add changelog

it is possible for a cutout to be missing; we need to allow for that.
Also, the hack I had before isn't needed when you appropriately query
for multiple wavelengths

remove some comments, still not done with the PR

minor correction

added basic docs

pep8 fixes.  The bracket closures DO match the indentation, though.
pep8 lies sometimes.

add local tests and cleanup some testing

more interactive tests revealed that my hacks didn't really work, this
commit has some fixes

move everything into astroquery.herschel.higal
