--- astroquery/higal/tests/setup_package.py
@@ -1,10 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-from __future__ import absolute_import
-
-import os
-
-
-def get_package_data():
-    paths = [os.path.join('data', '*.html'),
-             ]
-    return {'astroquery.higal.tests': paths}
